將兩腿深度防護提高至三倍
- 每腿必須具備下單量三倍的限價內可成交深度
- 五股訂單要求 Up 與 Down 各至少十五股深度
- 更新環境範例、說明及整數股數測試

+++ tests/test_polymarket_live_strategy.py
@@ -1206,7 +1206,9 @@ class LiveStrategyTests(unittest.IsolatedAsyncioTestCase):
             strategy.sim.SIM_DEPTH_CAP_FRACTION,
             1.0 / strategy.PAIR_MIN_DEPTH_MULTIPLIER,
         )
-        self.assertEqual(pos["shares"], depth * expected_fraction)
+        # 實際策略會把 outcome token 股數無條件捨去為整數；3x 深度時
+        # 40 / 3 = 13.333...，因此應送 13 股，不可期待小數股。
+        self.assertEqual(pos["shares"], float(int(depth * expected_fraction)))
 
     def _set_market_for_preflight(self):
         strategy.sim.state["market"] = {
